requirements.txt
Pillow~=10.0.0
requests~=2.31.0
Flask~=2.3.2